# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mpc-ext server",
    "description": "Holds one share of a 2-of-3 MPC key, joins signing, and verifies server-origin WebAuthn passkey ceremonies.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/v1/device-key": {
      "post": {
        "tags": [],
        "summary": "Registers a device key. Signing endpoints verify requests with this key.",
        "operationId": "register_device_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "device key registered"
          },
          "400": {
            "description": "malformed request"
          }
        }
      }
    },
    "/v1/dkg/round": {
      "post": {
        "tags": [],
        "summary": "Advances to the next round using the envelopes supplied.",
        "operationId": "advance_dkg",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceDkg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "round advanced, or DKG complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancedDkg"
                }
              }
            }
          },
          "404": {
            "description": "unknown or expired session"
          }
        }
      }
    },
    "/v1/dkg/session": {
      "post": {
        "tags": [],
        "summary": "Opens a DKG session and returns the server's round 1 envelopes.",
        "operationId": "start_dkg",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDkg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "session opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartedDkg"
                }
              }
            }
          },
          "400": {
            "description": "malformed request"
          }
        }
      }
    },
...
    "/v1/passkeys/assert/finish": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Verifies one passkey assertion and consumes its operation binding.",
        "operationId": "assert_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "assertion verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssertFinishResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/passkeys/assert/options": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Starts a passkey assertion bound to one exact operation.",
        "operationId": "assert_options",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "assertion options",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssertOptionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/passkeys/ceremony/status": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Returns browser ceremony progress to the extension without exposing credential state.",
        "operationId": "ceremony_status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CeremonyStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "browser ceremony status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CeremonyStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/passkeys/handoff": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Creates the one-use browser handoff after verifying device possession.",
        "operationId": "handoff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "browser ceremony handoff",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/passkeys/register/finish": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Verifies and stores a new passkey credential.",
        "operationId": "register_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "passkey registered"
          }
        }
      }
    },
    "/v1/passkeys/register/options": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Starts a passkey registration ceremony.",
        "operationId": "register_options",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "registration options",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOptionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sign/round": {
      "post": {
        "tags": [],
        "summary": "Advances the signing protocol by one round.",
        "operationId": "advance_sign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceSign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "round advanced, or signing complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancedSign"
                }
              }
            }
          },
          "404": {
            "description": "unknown or expired session"
          }
        }
      }
    },
    "/v1/sign/session": {
      "post": {
        "tags": [],
        "summary": "Opens a signing session and returns the server's round 1 envelopes.",
        "description": "The server is a second factor, so this is where policy belongs: rate limits, anomaly\nblocking and user confirmation all hang off this endpoint once authentication exists\n(`docs/server.md`).",
        "operationId": "start_sign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "session opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartedSign"
                }
              }
            }
          },
          "404": {
            "description": "no key for this wallet"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdvanceDkg": {
        "type": "object",
        "description": "A request to advance a DKG round.",
        "required": [
          "session_id",
          "wallet_id",
          "envelopes"
        ],
        "properties": {
          "envelopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WireEnvelope"
            },
            "description": "The envelopes being delivered to the server."
          },
          "session_id": {
            "type": "string",
            "description": "The session id, 64 hex characters."
          },
          "wallet_id": {
            "type": "string",
            "description": "The wallet identifier."
          }
        }
      },
...
      "AdvancedDkg": {
        "oneOf": [
          {
            "type": "object",
            "description": "Still running. Deliver the envelopes and call again.",
            "required": [
              "envelopes",
              "state"
            ],
            "properties": {
              "envelopes": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WireEnvelope"
                },
                "description": "The envelopes the server sends."
              },
              "state": {
                "type": "string",
                "enum": [
                  "inProgress"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "The DKG finished and the server stored share C.",
            "required": [
              "public_key",
              "state"
            ],
            "properties": {
              "public_key": {
                "type": "string",
                "description": "The SEC1 compressed public key, hex-encoded."
              },
              "state": {
                "type": "string",
                "enum": [
                  "completed"
                ]
              }
            }
          }
        ],
        "description": "The response to advancing a DKG round."
      },
...
      "AssertFinishRequest": {
        "type": "object",
        "description": "Completes an assertion ceremony.",
        "required": [
          "wallet_id",
          "challenge_id",
          "purpose",
          "operation_id",
          "digest",
          "credential"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "description": "The opaque challenge identifier returned by the options endpoint."
          },
          "credential": {
            "type": "object",
            "description": "The browser's `PublicKeyCredential` JSON object."
          },
          "digest": {
            "type": "string",
            "description": "Must exactly match the values bound when options were requested."
          },
          "operation_id": {
            "type": "string",
            "description": "Must exactly match the values bound when options were requested."
          },
          "purpose": {
            "type": "string",
            "description": "Must exactly match the values bound when options were requested."
          },
          "wallet_id": {
            "type": "string",
            "description": "The wallet identifier."
          }
        }
      },
      "AssertFinishResponse": {
        "type": "object",
        "description": "The result of a verified assertion.",
        "required": [
          "verified",
          "purpose",
          "operation_id"
        ],
        "properties": {
          "operation_id": {
            "type": "string",
            "description": "Echoes the already validated operation identifier."
          },
          "purpose": {
            "type": "string",
            "description": "Echoes the already validated policy purpose."
          },
          "verified": {
            "type": "boolean",
            "description": "Always true on a successful response."
          }
        }
      },
      "AssertOptionsRequest": {
        "type": "object",
        "description": "Starts an assertion ceremony bound to one exact operation.",
        "required": [
          "wallet_id",
          "purpose",
          "operation_id",
          "digest"
        ],
        "properties": {
          "digest": {
            "type": "string",
            "description": "The exact 32-byte operation digest, encoded as lowercase or uppercase hex."
          },
          "operation_id": {
            "type": "string",
            "description": "Application-generated operation identifier."
          },
          "purpose": {
            "type": "string",
            "description": "The policy operation, currently `sign` or `recovery`."
          },
          "wallet_id": {
            "type": "string",
            "description": "The wallet identifier."
          }
        }
      },
...
      "RegisterDeviceKey": {
        "type": "object",
        "description": "A device public-key registration request.",
        "required": [
          "wallet_id",
          "public_key"
        ],
        "properties": {
          "public_key": {
            "type": "string",
            "description": "The raw SEC1 public key, encoded as lowercase hex."
          },
          "wallet_id": {
            "type": "string",
            "description": "The wallet identifier."
          }
        }
      },
...
      "StartDkg": {
        "type": "object",
        "description": "A request to open a DKG session.",
        "required": [
          "wallet_id",
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string",
            "description": "The session id shared by every party, 64 hex characters."
          },
          "wallet_id": {
            "type": "string",
            "description": "The wallet identifier, generated by the client."
          }
        }
      },
...
      "StartedDkg": {
        "type": "object",
        "description": "The response to opening a DKG session.",
        "required": [
          "envelopes"
        ],
        "properties": {
          "envelopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WireEnvelope"
            },
            "description": "The round 1 envelopes the server sends."
          }
        }
      },
...
    }
  }
}